requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0
chromadb>=0.4.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pydantic>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
underthesea>=6.7.0
python-docx>=1.1.0
openpyxl>=3.1.0
pypdf>=3.17.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Additional dependencies for knowledge graph and semantic reasoning
sentence-transformers>=2.2.2
networkx>=3.0
scikit-learn>=1.3.0
plotly>=5.15.0
openai>=1.0.0  # Optional for advanced LLM features
numpy>=1.24.0
pandas>=2.0.0

# Neo4j and LLM dependencies
neo4j>=5.14.0
google-generativeai>=0.3.0